LPS-191765 portal-search-tuning-rankings-web: move _singleIndexToMultipleIndexImporter test logic and delete old tests
https://liferay.atlassian.net/browse/LPS-191765
