+++ modules/dxp/apps/portal-search-tuning/portal-search-tuning-rankings-web/src/test/java/com/liferay/portal/search/tuning/rankings/web/internal/index/creation/instance/lifecycle/RankingIndexPortalInstanceLifecycleListenerTest.java
@@ -10,6 +10,7 @@ import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.search.capabilities.SearchCapabilities;
 import com.liferay.portal.search.tuning.rankings.web.internal.index.RankingIndexCreator;
 import com.liferay.portal.search.tuning.rankings.web.internal.index.RankingIndexReader;
+import com.liferay.portal.search.tuning.rankings.web.internal.index.importer.SingleIndexToMultipleIndexImporter;
 import com.liferay.portal.search.tuning.rankings.web.internal.index.name.RankingIndexNameBuilder;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 
@@ -53,11 +54,30 @@ public class RankingIndexPortalInstanceLifecycleListenerTest {
 		ReflectionTestUtil.setFieldValue(
 			_rankingIndexPortalInstanceLifecycleListener, "_searchCapabilities",
 			_searchCapabilities);
+		ReflectionTestUtil.setFieldValue(
+			_rankingIndexPortalInstanceLifecycleListener,
+			"_singleIndexToMultipleIndexImporter",
+			_singleIndexToMultipleIndexImporter);
+	}
+
+	@Test
+	public void testActivatorSingleIndexToMultipleIndexImporterTrue()
+		throws Exception {
+
+		_setUpSingleIndexToMultipleIndexImporter(true);
+
+		_rankingIndexPortalInstanceLifecycleListener.portalInstanceRegistered(
+			Mockito.mock(Company.class));
+
+		Mockito.verify(
+			_singleIndexToMultipleIndexImporter, Mockito.times(1)
+		).importRankings();
 	}
 
 	@Test
 	public void testPortalInstanceRegisteredExistFalse() throws Exception {
 		_setUpRankingIndexReader(false);
+		_setUpSingleIndexToMultipleIndexImporter(false);
 
 		_rankingIndexPortalInstanceLifecycleListener.portalInstanceRegistered(
 			Mockito.mock(Company.class));
@@ -78,6 +98,7 @@ public class RankingIndexPortalInstanceLifecycleListenerTest {
 	@Test
 	public void testPortalInstanceRegisteredExistTrue() throws Exception {
 		_setUpRankingIndexReader(true);
+		_setUpSingleIndexToMultipleIndexImporter(false);
 
 		_rankingIndexPortalInstanceLifecycleListener.portalInstanceRegistered(
 			Mockito.mock(Company.class));
@@ -98,6 +119,7 @@ public class RankingIndexPortalInstanceLifecycleListenerTest {
 	@Test
 	public void testPortalInstanceUnregisteredExistsFalse() throws Exception {
 		_setUpRankingIndexReader(false);
+		_setUpSingleIndexToMultipleIndexImporter(false);
 
 		_rankingIndexPortalInstanceLifecycleListener.portalInstanceUnregistered(
 			Mockito.mock(Company.class));
@@ -118,6 +140,7 @@ public class RankingIndexPortalInstanceLifecycleListenerTest {
 	@Test
 	public void testPortalInstanceUnregisteredExistsTrue() throws Exception {
 		_setUpRankingIndexReader(true);
+		_setUpSingleIndexToMultipleIndexImporter(false);
 
 		_rankingIndexPortalInstanceLifecycleListener.portalInstanceUnregistered(
 			Mockito.mock(Company.class));
@@ -145,6 +168,14 @@ public class RankingIndexPortalInstanceLifecycleListenerTest {
 		);
 	}
 
+	private void _setUpSingleIndexToMultipleIndexImporter(boolean exist) {
+		Mockito.doReturn(
+			exist
+		).when(
+			_singleIndexToMultipleIndexImporter
+		).needImport();
+	}
+
 	private final RankingIndexCreator _rankingIndexCreator = Mockito.mock(
 		RankingIndexCreator.class);
 	private final RankingIndexNameBuilder _rankingIndexNameBuilder =
@@ -155,5 +186,8 @@ public class RankingIndexPortalInstanceLifecycleListenerTest {
 		RankingIndexReader.class);
 	private final SearchCapabilities _searchCapabilities = Mockito.mock(
 		SearchCapabilities.class);
+	private final SingleIndexToMultipleIndexImporter
+		_singleIndexToMultipleIndexImporter = Mockito.mock(
+			SingleIndexToMultipleIndexImporter.class);
 
 }
